Revert "Merge branch 'fix_gpp' into 'master'"
This reverts commit 0031dcbfca5656e5ecb335794102bbac6c603764, reversing
changes made to c9b58db405a3c4f95d598cf7b3b707189e10ccac.

+++ src/xc-threshold.c
@@ -98,373 +98,373 @@ typedef struct {
   double v4tau4[5];
 } xc_values_type;
 
-void index_lookup(size_t idx, char *legend) {
-  size_t offset = 0;
+void index_lookup(int idx, char *legend) {
+  int offset = 0;
   xc_dimensions dim;
   internal_counters_set_mgga(XC_POLARIZED, &dim);
   if (idx < offset + dim.rho) {
-    sprintf(legend, "rho[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "rho[%i]", idx - offset);
     return;
   }
   offset += dim.rho;
   if (idx < offset + dim.sigma) {
-    sprintf(legend, "sigma[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "sigma[%i]", idx - offset);
     return;
   }
   offset += dim.sigma;
   if (idx < offset + dim.lapl) {
-    sprintf(legend, "lapl[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "lapl[%i]", idx - offset);
     return;
   }
   offset += dim.lapl;
   if (idx < offset + dim.tau) {
-    sprintf(legend, "tau[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "tau[%i]", idx - offset);
     return;
   }
   offset += dim.tau;
 
   if (idx < offset + dim.vrho) {
-    sprintf(legend, "vrho[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "vrho[%i]", idx - offset);
     return;
   }
   offset += dim.vrho;
   if (idx < offset + dim.vsigma) {
-    sprintf(legend, "vsigma[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "vsigma[%i]", idx - offset);
     return;
   }
   offset += dim.vsigma;
   if (idx < offset + dim.vlapl) {
-    sprintf(legend, "vlapl[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "vlapl[%i]", idx - offset);
     return;
   }
   offset += dim.vlapl;
   if (idx < offset + dim.vtau) {
-    sprintf(legend, "vtau[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "vtau[%i]", idx - offset);
     return;
   }
   offset += dim.vtau;
   if (idx < offset + dim.v2rho2) {
-    sprintf(legend, "v2rho2[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v2rho2[%i]", idx - offset);
     return;
   }
   offset += dim.v2rho2;
   if (idx < offset + dim.v2rhosigma) {
-    sprintf(legend, "v2rhosigma[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v2rhosigma[%i]", idx - offset);
     return;
   }
   offset += dim.v2rhosigma;
   if (idx < offset + dim.v2sigma2) {
-    sprintf(legend, "v2sigma2[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v2sigma2[%i]", idx - offset);
     return;
   }
   offset += dim.v2sigma2;
   if (idx < offset + dim.v2rholapl) {
-    sprintf(legend, "v2rholapl[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v2rholapl[%i]", idx - offset);
     return;
   }
   offset += dim.v2rholapl;
   if (idx < offset + dim.v2sigmalapl) {
-    sprintf(legend, "v2sigmalapl[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v2sigmalapl[%i]", idx - offset);
     return;
   }
   offset += dim.v2sigmalapl;
   if (idx < offset + dim.v2lapl2) {
-    sprintf(legend, "v2lapl2[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v2lapl2[%i]", idx - offset);
     return;
   }
   offset += dim.v2lapl2;
   if (idx < offset + dim.v2rhotau) {
-    sprintf(legend, "v2rhotau[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v2rhotau[%i]", idx - offset);
     return;
   }
   offset += dim.v2rhotau;
   if (idx < offset + dim.v2sigmatau) {
-    sprintf(legend, "v2sigmatau[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v2sigmatau[%i]", idx - offset);
     return;
   }
   offset += dim.v2sigmatau;
   if (idx < offset + dim.v2lapltau) {
-    sprintf(legend, "v2lapltau[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v2lapltau[%i]", idx - offset);
     return;
   }
   offset += dim.v2lapltau;
   if (idx < offset + dim.v2tau2) {
-    sprintf(legend, "v2tau2[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v2tau2[%i]", idx - offset);
     return;
   }
   offset += dim.v2tau2;
   if (idx < offset + dim.v3rho3) {
-    sprintf(legend, "v3rho3[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v3rho3[%i]", idx - offset);
     return;
   }
   offset += dim.v3rho3;
   if (idx < offset + dim.v3rho2sigma) {
-    sprintf(legend, "v3rho2sigma[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v3rho2sigma[%i]", idx - offset);
     return;
   }
   offset += dim.v3rho2sigma;
   if (idx < offset + dim.v3rhosigma2) {
-    sprintf(legend, "v3rhosigma2[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v3rhosigma2[%i]", idx - offset);
     return;
   }
   offset += dim.v3rhosigma2;
   if (idx < offset + dim.v3sigma3) {
-    sprintf(legend, "v3sigma3[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v3sigma3[%i]", idx - offset);
     return;
   }
   offset += dim.v3sigma3;
   if (idx < offset + dim.v3rho2lapl) {
-    sprintf(legend, "v3rho2lapl[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v3rho2lapl[%i]", idx - offset);
     return;
   }
   offset += dim.v3rho2lapl;
   if (idx < offset + dim.v3rhosigmalapl) {
-    sprintf(legend, "v3rhosigmalapl[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v3rhosigmalapl[%i]", idx - offset);
     return;
   }
   offset += dim.v3rhosigmalapl;
   if (idx < offset + dim.v3sigma2lapl) {
-    sprintf(legend, "v3sigma2lapl[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v3sigma2lapl[%i]", idx - offset);
     return;
   }
   offset += dim.v3sigma2lapl;
   if (idx < offset + dim.v3rholapl2) {
-    sprintf(legend, "v3rholapl2[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v3rholapl2[%i]", idx - offset);
     return;
   }
   offset += dim.v3rholapl2;
   if (idx < offset + dim.v3sigmalapl2) {
-    sprintf(legend, "v3sigmalapl2[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v3sigmalapl2[%i]", idx - offset);
     return;
   }
   offset += dim.v3sigmalapl2;
   if (idx < offset + dim.v3lapl3) {
-    sprintf(legend, "v3lapl3[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v3lapl3[%i]", idx - offset);
     return;
   }
   offset += dim.v3lapl3;
   if (idx < offset + dim.v3rho2tau) {
-    sprintf(legend, "v3rho2tau[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v3rho2tau[%i]", idx - offset);
     return;
   }
   offset += dim.v3rho2tau;
   if (idx < offset + dim.v3rhosigmatau) {
-    sprintf(legend, "v3rhosigmatau[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v3rhosigmatau[%i]", idx - offset);
     return;
   }
   offset += dim.v3rhosigmatau;
   if (idx < offset + dim.v3sigma2tau) {
-    sprintf(legend, "v3sigma2tau[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v3sigma2tau[%i]", idx - offset);
     return;
   }
   offset += dim.v3sigma2tau;
   if (idx < offset + dim.v3rholapltau) {
-    sprintf(legend, "v3rholapltau[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v3rholapltau[%i]", idx - offset);
     return;
   }
   offset += dim.v3rholapltau;
   if (idx < offset + dim.v3sigmalapltau) {
-    sprintf(legend, "v3sigmalapltau[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v3sigmalapltau[%i]", idx - offset);
     return;
   }
   offset += dim.v3sigmalapltau;
   if (idx < offset + dim.v3lapl2tau) {
-    sprintf(legend, "v3lapl2tau[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v3lapl2tau[%i]", idx - offset);
     return;
   }
   offset += dim.v3lapl2tau;
   if (idx < offset + dim.v3rhotau2) {
-    sprintf(legend, "v3rhotau2[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v3rhotau2[%i]", idx - offset);
     return;
   }
   offset += dim.v3rhotau2;
   if (idx < offset + dim.v3sigmatau2) {
-    sprintf(legend, "v3sigmatau2[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v3sigmatau2[%i]", idx - offset);
     return;
   }
   offset += dim.v3sigmatau2;
   if (idx < offset + dim.v3lapltau2) {
-    sprintf(legend, "v3lapltau2[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v3lapltau2[%i]", idx - offset);
     return;
   }
   offset += dim.v3lapltau2;
   if (idx < offset + dim.v3tau3) {
-    sprintf(legend, "v3tau3[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v3tau3[%i]", idx - offset);
     return;
   }
   offset += dim.v3tau3;
   if (idx < offset + dim.v4rho4) {
-    sprintf(legend, "v4rho4[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4rho4[%i]", idx - offset);
     return;
   }
   offset += dim.v4rho4;
   if (idx < offset + dim.v4rho3sigma) {
-    sprintf(legend, "v4rho3sigma[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4rho3sigma[%i]", idx - offset);
     return;
   }
   offset += dim.v4rho3sigma;
   if (idx < offset + dim.v4rho2sigma2) {
-    sprintf(legend, "v4rho2sigma2[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4rho2sigma2[%i]", idx - offset);
     return;
   }
   offset += dim.v4rho2sigma2;
   if (idx < offset + dim.v4rhosigma3) {
-    sprintf(legend, "v4rhosigma3[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4rhosigma3[%i]", idx - offset);
     return;
   }
   offset += dim.v4rhosigma3;
   if (idx < offset + dim.v4sigma4) {
-    sprintf(legend, "v4sigma4[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4sigma4[%i]", idx - offset);
     return;
   }
   offset += dim.v4sigma4;
   if (idx < offset + dim.v4rho3lapl) {
-    sprintf(legend, "v4rho3lapl[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4rho3lapl[%i]", idx - offset);
     return;
   }
   offset += dim.v4rho3lapl;
   if (idx < offset + dim.v4rho2sigmalapl) {
-    sprintf(legend, "v4rho2sigmalapl[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4rho2sigmalapl[%i]", idx - offset);
     return;
   }
   offset += dim.v4rho2sigmalapl;
   if (idx < offset + dim.v4rhosigma2lapl) {
-    sprintf(legend, "v4rhosigma2lapl[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4rhosigma2lapl[%i]", idx - offset);
     return;
   }
   offset += dim.v4rhosigma2lapl;
   if (idx < offset + dim.v4sigma3lapl) {
-    sprintf(legend, "v4sigma3lapl[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4sigma3lapl[%i]", idx - offset);
     return;
   }
   offset += dim.v4sigma3lapl;
   if (idx < offset + dim.v4rho2lapl2) {
-    sprintf(legend, "v4rho2lapl2[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4rho2lapl2[%i]", idx - offset);
     return;
   }
   offset += dim.v4rho2lapl2;
   if (idx < offset + dim.v4rhosigmalapl2) {
-    sprintf(legend, "v4rhosigmalapl2[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4rhosigmalapl2[%i]", idx - offset);
     return;
   }
   offset += dim.v4rhosigmalapl2;
   if (idx < offset + dim.v4sigma2lapl2) {
-    sprintf(legend, "v4sigma2lapl2[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4sigma2lapl2[%i]", idx - offset);
     return;
   }
   offset += dim.v4sigma2lapl2;
   if (idx < offset + dim.v4rholapl3) {
-    sprintf(legend, "v4rholapl3[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4rholapl3[%i]", idx - offset);
     return;
   }
   offset += dim.v4rholapl3;
   if (idx < offset + dim.v4sigmalapl3) {
-    sprintf(legend, "v4sigmalapl3[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4sigmalapl3[%i]", idx - offset);
     return;
   }
   offset += dim.v4sigmalapl3;
   if (idx < offset + dim.v4lapl4) {
-    sprintf(legend, "v4lapl4[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4lapl4[%i]", idx - offset);
     return;
   }
   offset += dim.v4lapl4;
   if (idx < offset + dim.v4rho3tau) {
-    sprintf(legend, "v4rho3tau[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4rho3tau[%i]", idx - offset);
     return;
   }
   offset += dim.v4rho3tau;
   if (idx < offset + dim.v4rho2sigmatau) {
-    sprintf(legend, "v4rho2sigmatau[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4rho2sigmatau[%i]", idx - offset);
     return;
   }
   offset += dim.v4rho2sigmatau;
   if (idx < offset + dim.v4rhosigma2tau) {
-    sprintf(legend, "v4rhosigma2tau[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4rhosigma2tau[%i]", idx - offset);
     return;
   }
   offset += dim.v4rhosigma2tau;
   if (idx < offset + dim.v4sigma3tau) {
-    sprintf(legend, "v4sigma3tau[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4sigma3tau[%i]", idx - offset);
     return;
   }
   offset += dim.v4sigma3tau;
   if (idx < offset + dim.v4rho2lapltau) {
-    sprintf(legend, "v4rho2lapltau[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4rho2lapltau[%i]", idx - offset);
     return;
   }
   offset += dim.v4rho2lapltau;
   if (idx < offset + dim.v4rhosigmalapltau) {
-    sprintf(legend, "v4rhosigmalapltau[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4rhosigmalapltau[%i]", idx - offset);
     return;
   }
   offset += dim.v4rhosigmalapltau;
   if (idx < offset + dim.v4sigma2lapltau) {
-    sprintf(legend, "v4sigma2lapltau[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4sigma2lapltau[%i]", idx - offset);
     return;
   }
   offset += dim.v4sigma2lapltau;
   if (idx < offset + dim.v4rholapl2tau) {
-    sprintf(legend, "v4rholapl2tau[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4rholapl2tau[%i]", idx - offset);
     return;
   }
   offset += dim.v4rholapl2tau;
   if (idx < offset + dim.v4sigmalapl2tau) {
-    sprintf(legend, "v4sigmalapl2tau[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4sigmalapl2tau[%i]", idx - offset);
     return;
   }
   offset += dim.v4sigmalapl2tau;
   if (idx < offset + dim.v4lapl3tau) {
-    sprintf(legend, "v4lapl3tau[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4lapl3tau[%i]", idx - offset);
     return;
   }
   offset += dim.v4lapl3tau;
   if (idx < offset + dim.v4rho2tau2) {
-    sprintf(legend, "v4rho2tau2[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4rho2tau2[%i]", idx - offset);
     return;
   }
   offset += dim.v4rho2tau2;
   if (idx < offset + dim.v4rhosigmatau2) {
-    sprintf(legend, "v4rhosigmatau2[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4rhosigmatau2[%i]", idx - offset);
     return;
   }
   offset += dim.v4rhosigmatau2;
   if (idx < offset + dim.v4sigma2tau2) {
-    sprintf(legend, "v4sigma2tau2[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4sigma2tau2[%i]", idx - offset);
     return;
   }
   offset += dim.v4sigma2tau2;
   if (idx < offset + dim.v4rholapltau2) {
-    sprintf(legend, "v4rholapltau2[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4rholapltau2[%i]", idx - offset);
     return;
   }
   offset += dim.v4rholapltau2;
   if (idx < offset + dim.v4sigmalapltau2) {
-    sprintf(legend, "v4sigmalapltau2[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4sigmalapltau2[%i]", idx - offset);
     return;
   }
   offset += dim.v4sigmalapltau2;
   if (idx < offset + dim.v4lapl2tau2) {
-    sprintf(legend, "v4lapl2tau2[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4lapl2tau2[%i]", idx - offset);
     return;
   }
   offset += dim.v4lapl2tau2;
   if (idx < offset + dim.v4rhotau3) {
-    sprintf(legend, "v4rhotau3[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4rhotau3[%i]", idx - offset);
     return;
   }
   offset += dim.v4rhotau3;
   if (idx < offset + dim.v4sigmatau3) {
-    sprintf(legend, "v4sigmatau3[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4sigmatau3[%i]", idx - offset);
     return;
   }
   offset += dim.v4sigmatau3;
   if (idx < offset + dim.v4lapltau3) {
-    sprintf(legend, "v4lapltau3[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4lapltau3[%i]", idx - offset);
     return;
   }
   offset += dim.v4lapltau3;
   if (idx < offset + dim.v4tau4) {
-    sprintf(legend, "v4tau4[%lu]", (unsigned long)(idx - offset));
+    sprintf(legend, "v4tau4[%i]", idx - offset);
     return;
   }
   offset += dim.v4tau4;
@@ -805,7 +805,7 @@ size_t check_xc(int id, int nspin, xc_values_type values, double threshold) {
 
   /* Check the output for NaNs */
   double *ptr = (double *)&values;
-  for (size_t i = 0; i < sizeof(xc_values_type) / sizeof(double); i++)
+  for (int i = 0; i < sizeof(xc_values_type) / sizeof(double); i++)
     /* We have encountered an infinity or NaN */
     if (!isfinite(ptr[i])) {
 #ifdef XC_DEBUG
